Extract japanese_study_words_algo reporting into study_words_report
Same treatment as kanji_vocab_report: move print_report (~250 lines of stdout
stats — tier/length/kanji-count/POS/JLPT breakdowns, textbook overlap, anomaly
lists) into a sibling study_words_report module. Diagnostic-only, writes no file,
pure relocation with no effect on shipped output.

Constants/helpers the report needs (OVERRIDE_TAG, OUTPUT_TEXTBOOK_TAG,
read_freq_rows, COL_WORD/COL_JLPT) stay on japanese_study_words_algo and are read
through an `sw` alias; the report is imported late inside main() to avoid an import
cycle. Also drops the now-unused Counter and classify_pos imports from the algo.

japanese_study_words_algo.py: 809 -> 558 lines.

Verified: byte-compile, no dangling references, cross-module imports (kanji_load's
study-words attributes intact), and an offline harness that runs the extracted
print_report end-to-end with a stub resolver (all sections render).
